+++ rerun_py/rerun_sdk/rerun/sinks.py
@@ -9,7 +9,7 @@ import rerun_bindings as bindings  # type: ignore[attr-defined]
 from rerun.blueprint.api import BlueprintLike, create_in_memory_blueprint
 from rerun.recording_stream import RecordingStream, get_application_id
 
-from .spawn import _spawn_viewer
+from ._spawn import _spawn_viewer
 
 # --- Sinks ---
